refactor(queue): extract shared types and mappers to lib/queue
